Clean up examples
